# pb/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bidi/stream/example/echo": {
      "post": {
        "operationId": "EchoBiDiStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/pbStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    },
    "/v1/client/stream/example/echo": {
      "post": {
        "operationId": "EchoClientStream",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    },
    "/v1/example/echo": {
      "post": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    },
    "/v1/example/ping": {
      "get": {
        "operationId": "Ping",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbPongMessage"
            }
          }
        },
        "tags": [
          "YourService"
        ]
      }
    },
    "/v1/server/stream/example/echo": {
      "post": {
        "operationId": "EchoServerStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/pbStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbStringMessage"
            }
          }
        ],
        "tags": [
          "YourService"
        ]
      }
    }
  },
  "definitions": {
    "pbPongMessage": {
      "type": "object",
      "properties": {
        "ip": {
          "type": "string"
        }
      }
    },
    "pbStringMessage": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "http_status": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "x-stream-definitions": {
    "pbStringMessage": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/pbStringMessage"
        },
        "error": {
          "$ref": "#/definitions/runtimeStreamError"
        }
      },
      "title": "Stream result of pbStringMessage"
    }
  }
}